feat : event listener class 추가 - facade는 협업의 공간 및 단일 제어판으로만 활용, 이벤트를 listen하는 책임을 가진 별도의 class추가
